intrin -> runtime; util is mainly in charge of compilation time
